Create 이미지리스트,이미지삭제

+++ server/controllers/images/imageDelete.js
@@ -1,13 +1,25 @@
 const { posts_imgs } = require("../../models");
+//const { isAuthorized } = require("../tokenFunctions");
 
 module.exports = async (req, res) => {
-  const userinfo = req.params.id;
+  const userImg = req.params.id;
 
-  if (!userinfo) {
-    res.satus(404).send({ message: "Not found" });
+  if (!userImg) {
+    return res.satus(404).send({ message: "Not found" });
+  } else {
+    await posts_imgs.destroy({
+      where: { id: userImg },
+    });
+    return res.status(200).send({ message: "successfully deleted img" });
   }
-  await posts_imgs.destroy({
-    where: { id: userinfo },
-  });
-  return res.status(200).send({ message: "successfully deleted img" });
 };
+// const userImg = req.params.id;
+
+// if (!isAuthorized) {
+//   return res.satus(404).send({ message: "Not found" });
+// } else {
+//   await posts_imgs.destroy({
+//     where: { id: userImg },
+//   });
+//   return res.status(200).send({ message: "successfully deleted img" });
+// }
